Created the MLBPlayer class & RetrievePlayerInfo
MLBPlayer class stores all stat info for a player, including the box score for a specified game and their season stats. It will soon be used to calculate an “expected fantasy points / game” using an algorithm that takes an MLBPlayer. RetrievePlayerInfo creates MLBPlayers for every player for a specific game and populates them with the gamelong and season stats.