Created IAlgorithm and Implemented 2 Algorithms; Created Correlation Tests
Created IAlgorithm which defines an algorithm based off of a Fantasy Player
Created two MLBPlayer Algorithm. The first one which ranks the player based off of their batting average and ERA. The second which ranks the players randomly.
Used KendallTau rank test to assess the correlation between a predicted list of ranks and the actual list of ranked players based on their fantasy points.